src/engines/skyvern/server/requirements.txt
skyvern>=0.1.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
